rune_util: refactor to not double-convert to/from uint32/rune
remove duplicate conversions between rune type and uint_32_t
(called 'runelike' in the source for want of a better term).
Kept `runefold` func because FilterFunc (used in DFA) expects
a instead of uint32 and didn't want to change that.

Also fixed naming convention in rune_util: `__` prefix and `static`
modifier for stuff specific to the file, no prefix for stuff used
outside the file. Didn't see a reason to use __ for stuff used elsewhere.