AddDocumentCtx: Remove context information
AddDocumentCtx does not need its own ThreadSafeContext. It's only needed
when accessing the DB (in which case it can be generated on demand) or
when sending a reply (in which case we use the BlockedClient APIs).

I've also modified the API a bit so that the blocking and flow is more
abstracted from the caller.