Move PARTIAL logic to document.c
This moves all the logic for partial updates to document.c with minimal
API changes in AddDocumentCommand.

If PARTIAL is detected, the handling chain is short-circuited and the
updates are done within the main thread.

WIP - I might want to still do this in its own thread, depending on the
size?