array: Allow to use custom allocators
If we know the kind of allocator needs to be used in the end, it's
better to do it now rather than copy it over because we didn't use the
right allocator. This does add more space into the Array struct, but
this is acceptable since Array is not meant to be stored anywhere (it
already has a `cap` field), but rather used as a convenient way to grow
a buffer.