Ensure that faulty term length estimations don't crash highlighter
We use whitespace tokenization by default when determining the length of
the highlighted term. This doesn't work well in Chinese, and users
should really use `LANGUAGE chinese` when searching for such documents.
However in the event that the `LANGUAGE` directive is not use, we might
end up in a situation where the length of the "current" term overlaps
the length of the next term (see tests where we search for "redis").

In this case, we don't want to crash.