Convert traditional chinese to simplified for tokenizing
This allows traditional chinese to be converted to simplified Chinese
for the purposes of tokenization. The fix is implemented by actually
having Friso (our dictionary library) convert the chinese characters to
their simplified variants. This means that the inverted indexes contain
simplified words rather than traditional. The original document remains
the same however.

The mapping between the characters can be regenerated using the included
python script