Dmd improvements (#242)
* Allow to define `RS_NO_U128` to avoid __uint128_t

My IDE doesn't seem happy with it. This might also be a good workaround
to compare performance, memory etc.

* Allow "binary" keys to be stored in document table

This allows "binary" keys (aka keys with embedded NUL bytes) to be
stored in the doc table. Summary of the changes:

- DocTable API now requires a length argument with any key argument
- DocumentMetaData stores the key as an `sds` rather than a `char*`
- RDB format changed slightly (version bumped). Terminal NUL byte is no
  longer stored. We also can't directly use the rm_malloc/rm_free when
  loading the string, either.
- Added tests - both unit and pytests

* Add RSDocumentKey wrapper type

This type wraps/abstracts the key/keyLen for the doc table metadata

* CFLAGS: Missing function declarations should be compile errors

* Replace DocTable*R functions with a single MakeDocKeyR

To convert RedisModuleString to DocumentKey

* Revise DMD key helper functions

* RSDocumentKey member renaming

Remove `key` prefix

* Hide sds from redisearch.h

* DMD: Documentation and boilerplate fixes