(De)Serialize sortables when rewriting AOF
We didn't previously serialize sortables when rewriting AOF, but still
set the 'sortable' flag on the document, which meant that code that
checked for the presence of the flag but didn't check the pointer would
dereference inaccessible memory.

This fix serializes the sorting vector extending the internal DTADD
command used for AOF replication.