Please make the build reproducible
Whilst working on the Reproducible Builds effort [0], we noticed
that redisearch could not be built reproducibly. This is because
it links (etc.) in filesystem ordering which is non-determinstic.

Adding some $(sort ...) calls fixes this for me. You can test this
by using our "disorderfs" tool to mangle the directory ordering.

 [0] https://reproducible-builds.org/