Allow to disable partial return timeout
`ON_TIMEOUT [RETURN|FAIL]` can be set at the module level. When set, the
module will try to refrain from delivering partial results to the user
in case a query times out.

This also introduces the framework to allow independent query-level
timeout policies and intervals, but we need to add that to the command
parser as well.