Document table compaction
The doc table is now a arrays of linked lists of dmds, its operate as a simple hash table.
The hash is a simple modulo operation on the document id.
A new document is added to the tail of the relavent linked list.
The dmds is managed using ref counts to make sure we will not free the dmd when some other
query thread is using it.