Clean sortable handling
This makes the sortables be a single consistent block of
RS_SORTABLE_MAX. This shouldn't make a big difference (several bytes
here and there; as long as RS_SORTABLE_MAX remains within a reasonable
size), and allows us to extend the sortables field size. It also has the
benefit (and this is why I did this in the first place) of ensuring that
the sortable memory block remains the same throughout the lifetime of
the spec (since it seems there are things which reference it).