fixed: redisearch crashes on gc invocation after deleting and recreating an index
When deleting and index, if the gc thread is inside the blocking stage:
if ((rc = pthread_cond_timedwait(&tm->cond, &tm->lock, &timeout)) == ETIMEDOUT) {
then it will get the signal and exist. But if the gc is currently running then it will not get the signal and will keep running.
Now as long as we do not create an index with the same name we are safe, the gc will failed to open it and will do nothing.
The moment we recreate the index, the gc reopen it and start running on garbage …
The more docs we have the greater the probability that we send the signal while the gc is running.

The solution is set a unique id to each index on index creation, when gc iteration starts we check that the unique id match.
If the unique id is not match we just drop the gc thread.