Don't make phonetic search the default
This has the potential to confuse users and slow things down, if
phonetic search is actually enabled. Rather, disable phonetic search
unless explicitly enabled via {$phonetic:true} as a query attribute.

We can be smarter about this once we detect the given field the
attribute applies to, however we can't know this at the query parse
stage yet.