Remove optimization which causes crash (#484)
We tried using the parent string as a means to be able to split the
original string without copying for each split. However, this is
problematic because we have no way of communicating or setting the state
of the parent string. In other words, we have no way of communicating to
the parent RSValue that the string should not be modified (technically..
it might be possible..)