see #491, add code_style.sh script that perform style changes on modified files
It is possible to run this script with --dry-run option, in this case the script
will only verify styling without changing anything.
It is possible to run this script with --install-pre-commit-script, in this case
the script will install a git pre commit hook which will not allow to commit
as long as there is a styling errors.