SendReply: Allow to send typed numeric/string values
In the Great Refactor, we removed the WITHSCHEMA option for internal
usage. The WITHSCHEMA option was fairly invasive and complex. However,
Redis still lacks a way to natively distinguish between floating point
and string values on the wire. As a hack, I've made it so that numeric
values are transmitted as "status" messages in-line with the responses.
Because of how numbers are encoded at the protocol level, this is
essentially a 'free' change.

This is an internal mechanism only, and is activated by using the
`_NUM_SSTRING` option on the request.