Split pipeline building into separate functions
Building a pipeline involves three stages; the first is establishing the
"source" pipeline, then applying any filters, and finally performing
things like sorting and paging on the results.

The BUILDPIPELINE_NO_ROOT option (new) will avoid building the root
phase of the pipeline. In this case, the user needs to supply his own
root, and manually attach it.