Implement multi-type indexing
This allows the same field to be indexed in different ways depending on
user input. This is still WIP because it's missing a proper interface.

This commit also significantly overhauls the way fields are handled
because a single field can carry multiple "types".

Currently, multitype fields cannot be sortable due to the ambiguity of
the sorting preference (e.g. is `1` lexically sorted or numerically so?)