Allow to build redisearch as a static library
Use -DRS_BUILD_STATIC=ON in the CMake command line to enable.

The API symbols are RediSearch_XXX for both extern and static modes. If
you are a Redis module wishing to use Redisearch's APIs via
ExportApi/GetApi mechanism within Redis, ensure that the preprocessor
macro -DREDISEARCH_API_EXTERN is enabled, and ensure that
REDISEARCH_INIT_SYMBOLS and RediSearch_Initialize() are called.

Note that INIT_SYMBOLS and Initialize() can both be used in static lib
mode as well, except that in those cases they are no-ops