Fix leaks and refactor Document functions
This completely refactors the functions related to the Document object,
to have more clear semantics regarding ownership of the contained
strings. The main change is removing the dependency of AutoMemory to
function correctly, and instead reverting to manual memory management.

AutoMemory can potentially slow things down if there are a lot of
pointers which are being added, and some other pointers being freed;
since Redis needs to traverse a linked list whenever freeing a pointer
in order to 'remove' the pointer from the automemory list

This also removes some unsued Document_ functions