Don't return from sortables in aggregation
The fields in sortable are not an exact duplicate of the fields as
stored within the document. Therefore, we should never return from them,
since it will yield inconsistent data with normal search results.

Some more changes were done here in order to compensate for the fact
that typing information is needed to be added internal to items fetched
from the Redis hash