Parser text expr (#2534)
This commit splits the parsing token expr into expr and text_expr.

Before this change, expr was used to hold a parsed part of the query, but also to hold a parsed text expression ("hello world | war" for example), and this lead to allowing queries like "@title:@year:[2000 2010]" which make no sense (the @year:[2000 2010] was first reduced into expr, and then @title:expr rule was used to parse the rest).

This PR introduces text_expr token, which is similar to expr but holds only parsed text expressions. every rule that used expr was updated to work with the proper token it was supposed to use, and the few rules that made sense both ways were duplicated for each token.