Avoid terms dict rehashing on multiple readers.
On LLAPI, we allow multiple readers to simultaniusly read from the index.
In such case we can not allow rehashing the term dictionary because rehashing
is not thread safe.

The PR add the ability to disable rehashing on dictionary and use it from LLAPI
when locking the index for reads.