Fix RED-86036, improve performance on ID's list iterator.
When INKEYS option is used along side a wildcard query, the wildcard query is intersected
with ID's list iterator that contains a list of requested docId's.

The `IL_SkipTo` function (of the ID's list iterator) was malfunction, in some cases the binary search would
have return a docId which is less than the given docId that was asked to skip to.
This causes the intersect iterator to travers on all the document returned from the wildcard query
instead of skipping them. By fixing the binary search, the performance was drastically improve.

We can see it using the test that was added to the PR. The test uses `FT.PROFILE` API to check how many times
the `IL_SkipTo` was called. Without the fix, we call this function 1000 times (one for each document in the index).
After the fix, the function are called 2 times (one for each document in the ID's list).