building lookup keys required for sorting during building the pipeline, skipping fields that are already loaded
Unless its a json index, we use the loader only for keys that were not loaded in previoud steps.

Find where to place buffer-locker and unlockerin the pipeline by iterating the final pipeline.
If there's a pager the unlocker is placed before it (the pager is the unlocker upstream) because the pager declares EOF.

remove unused rlookup flag - RLOOKUP_F_OUTPUT

added RLOOKUP_F_ORIGINAL_VALUE_DOCSRC and RLOOKUP_F_ISLOADED flags

added flags field and RESULT_PROCESSOR_F_ACCESS_REDIS flag the rp base flags

added a function to add the sorter to list of fields to load. The list is intiated during building the pipeline and not during the search itself.