added a pointer to the spec's strong ref in the spec for quick access
changed global cursors list name to g_CursorsList
AREQ_StartCursor instead of the spec's name, it  gets a strongref to the spec to initialize a cursor with a wak reference to the spec.
It is no longer the spec's responsability to free the its cursor when it is dropped.

Cursors_Reserve -  instead of the spec's name, it  gets a strongref to the spec
Cursors_RenderStats,Cursors_RenderStatsForInfogets a spec pointer instead of a name.
removed Cursors_PurgeWithName (was used by the spec to clean up its cursors)
cursor list - removed specinfo array,
cursor - removed pointer to the global lost
removed CursorList_Empty from Indexes_Free -TODO: add destroy after all cursor executions is done