Make cursors thread safe (#3620)
* remove unused function from coordinator -addIndexCursor()
used cursors and cursors capacity is saved in the spec instead of in the cursors list
it is initialized upon specs creation
so we don't need to remove the spec from the cursors list when we drop the index

* added a pointer to the spec's strong ref in the spec for quick access

changed global cursors list name to g_CursorsList
AREQ_StartCursor instead of the spec's name, it  gets a strongref to the spec to initialize a cursor with a wak reference to the spec.
It is no longer the spec's responsability to free the its cursor when it is dropped.

Cursors_Reserve -  instead of the spec's name, it  gets a strongref to the spec
Cursors_RenderStats,Cursors_RenderStatsForInfogets a spec pointer instead of a name.

removed Cursors_PurgeWithName (was used by the spec to clean up its cursors)
cursor list - removed specinfo array,
cursor - removed pointer to the global list

* added test that drops the index while the cursor is idle

* if we are in coordinator we send a strong ref that points to null instead of a valid spec

added testExceedCursorCapacity + BG
testIndexDropWhileIdleBG

nullify strong ref in index free