Fix tls on Redis 7.2 oss cluster
On Redis 7.2 (specifically on 7.2-rc3) there was a change
made that allow running on cluster with both, tls and none
tls port.
As part of this change, `cluster slot` command will return the
port base of the connection type (tls or not tls). A side effect
of this change is that when `cluster slot` called using `RM_Call`
it is always gets the none tls port (which break our support of tls).

To fix this, use `RedisModule_GetClusterNodeInfo` Redis module API to
get the correct port.

Notice that this is a minimal fix, we should extend the Redis module API
to provide us with all the information we need (including slot range) so
we will not have to use `cluster slot` command at all.

(cherry picked from commit 59c6636e379208dff7c250e78ed66c3691e4df1d)