WIP: fix numeric index stats
added comments to the gc flow of the numeric index.

IndexBlock_Repair(): removed saving oldFirstBlock as we anyway delete empty blocks.

Fix the root maxdepth that was never updated since the root in not rebalanced.

FGC_childRepairInvidx(): `ixmsg.lastblkBytesCollected` contained the entire bytes collected in block. If the last block is skipped (due to being modified by the parent while the fork was running), we discard the collected changes. Means, we need to update the general collection info. because  `ixmsg.lastblkBytesCollected` contained all the blocks bytes, the general info became 0.
checkLastBlock(): in case we discard the changes made in the last block, decrease `info->nentriesCollected` by  info->lastblkEntriesRemoved;

wip: added more information, including block efficiency to the debug output of several functions. still need to be tested