Added unit tests for inverted index gc scenarios.
FGC_applyInvertedIndex():
in case all the blocks are deleted, we can conclude the new index size by substracting the `idxData->numDelBlocks` recieved from the child from the current idx size.
Since the current index size is always equal or bigger than the one the fork handeled, it is safe to use the idx->blocks memory. If the last block was skipped it won't be included in `idxData->numDelBlocks`.