MOD-5757 fix inverted index GC related bugs and numeric index root max depth (#3874)
added comments to the gc flow of the numeric index.

IndexBlock_Repair(): removed saving oldFirstBlock as we anyway delete empty blocks.

Fix the root maxdepth that was never updated since the root in not rebalanced.

FGC_childRepairInvidx(): `ixmsg.lastblkBytesCollected` contained the entire bytes collected in block. If the last block is skipped (due to being modified by the parent while the fork was running), we discard the collected changes. Means, we need to update the general collection info. because  `ixmsg.lastblkBytesCollected` contained all the blocks bytes, the general info became 0.
checkLastBlock(): in case we discard the changes made in the last block, decrease `info->nentriesCollected` by  info->lastblkEntriesRemoved;

* Added unit tests for inverted index gc scenarios.

FGC_applyInvertedIndex():
in case all the blocks are deleted, we can conclude the new index size by substracting the `idxData->numDelBlocks` recieved from the child from the current idx size.
Since the current index size is always equal or bigger than the one the fork handeled, it is safe to use the idx->blocks memory. If the last block was skipped it won't be included in `idxData->numDelBlocks`.


changed testNumEntries name to testOverrides, make sure that the tree doesn't change after the first build.

Aligned testSanity expected output with the new 'FT.DEBUG', 'numidx_summary' command's output

in testEmptyNumericLeakIncrease, assert that numentries is exactly as expected. (after the gc is unvoked it should be equal to the number of docs)

added numeric_tree_summary function to common

debug_commands.c:
added POSTPONED_LEN_ARRAY macro to make nested array more easy to follow.

DocTable_Exists(): return immediately if we found the document. return 1 if it's valid and 0 if it's marked as deleted.