Fix uniquesum and cardinalities values corruption after gc - MOD-5815 (#3892)
* Save the hash table key as FP binary representation of the cardinality value.

* added more values to the test

* Fixed CardinalityValue array len

in recvCardvals, we initialize arrayof(CardinalityValue) with a **capacity** of len (cardinality values count) * sizeof(CardinalityValue). Although capacity should be in elements and not in bytes.
then we read the cardinality values sent from the child directly to the memory allocated for the array, instead of using array_append. It is probably more efficient and quick.
However, As a result, the array header is not updated with the correct length. The array length is then assigned to the range cardinality. So before this fix, is was zero.
Since we know in advance to exact number of the cardinality values, we can safly use array_newlen to initialize the CardinalityValue array. This way the array is initialized with the correct CardinalityValue count.

Added split test that checks split decision efficiency and correctness.

add comment to testUniqueSum

added timeout as optional argument to GC_FORCEINVOKE to use it during debugging to avoid terminating the process .

* check that the number of args for GCForceInvoke is 1 or 2

* minor comments