add optional free callback to coordinator requests.
Actually we don't need  a free cb if there's a client waiting for the request response.

MR_UpdateTopology generates a periodic request which might be still in the queue on shutdown, even on a proper run and the sanitizer will raise a leak error.
The free cb frees the topology and the request itself.

added ctx to MR_Destroy and Coordinator_CleanupModule to print logs

changed queueItem->cb type to MRQueueCallback