MOD-6019 fix flaky test_followhashes:testManyPrefixes (#4189)
* The original purpose of the testManyPrefixes test was to measure the performance enhancement introduced in PR #2560. (enhanced performance of the cleanup process upon FLUSHALL by optimizing the way the spec:prefixes dictionary and the cursors lists are cleared)
However, the test was unstable and could timeout on some platforms, even after the optimization. The sanitizer was also disabled, which is not ideal for a cleanup test.

The PR then modifies the test to focus on the functionality of the cleanup process, rather than its performance. This change ensures that the test is more stable and that it properly checks that the cleanup process is working as expected, with the sanitizer enabled.

* added a debug command to print the prefix trie basic information

chnaged testDeleteManyPrefixes name to testFlushallManyPrefixes
checks that the trie is empty using the new debug command

* remove new lines

* add the new debug command to test_debug_commands:TestDebugCommands.testDebugHelp

* fix

* cluster fix