This PR fixes MOD-6281
The sanitizer reports read after writes of the cluster resources such as the nodes trie and the nodes themselves.

The sequence of events leading to the leak:
1. `rqAsyncCb` pulls a job from the queue
2. shutdown even is fired
3. all pending jobs within the queue are removed
4. cluster resources are realesed
5. `rqAsyncCb` invokes a callback that involves accessing to cluster resources that have already been released

`rqAsyncCb`: the uv callback invoked from a background thread. in each invokation of this cb, we pull and execute all pending jobs from the queue.
Most uv requests are pushed from the main thread, with the exception of `uvUpdateTopologyRequest`, triggered periodically (every 1 second) according to the following path:
rmutilTimer_Loop->updateTopoCB->RefreshClusterCommand->MR_UpdateTopology->RQ_Push(uvUpdateTopologyRequest)

Note:
* rmutilTimer_Loop operates in the background.
* updateTopoCB locks the GIL, unlocking it after pushing to the queue.
* MR_Destroy() is responsible for freeing the queue, closing UV, and executing cluster cleanups.
One of the assumptions made here is the exclusive push of requests to rq either from the main thread or under GIL protection, ensuring no incoming requests during MR_Destroy() execution.

However, the return of uv_close doesn't guarantee immediate closure of the UV thread. This led to a race condition: rqAsyncCb manages to pull a job from the queue before RQ_Free is called, and subsequently, MR_Destroy() releases cluster resources while rqAsyncCb invokes a callback that might access those released structures.

This commit contains the reproduction of the bug.
we use the 'order_for_debug' global variable to ensure the commands acure in the order that causes the leak.
The scenario goes as follows:
1. rqAsyncCb is paused before popping from the queue, ensuring it remains active during shutdown. test_coordinator:testBugFixes includes sleeping for 2 seconds to make sure at least one periodic request is was pushed to the queue before shutdown is fired.
2. The shutdown includes calling to MR_Destroy(), that updates `order_for_debug` variable to enable rqAsyncCb to pull a job from the queue, and trigger RQ_Free().
3. To precisely replicate the reported leak, we wait for the release of the cluster resources in a specific location of the request cb. in this case: uvUpdateTopologyRequest->MRCLuster_UpdateTopology->_MRClsuter_UpdateNodes. we wait on order_for_debug before `TrieMap_Delete()`.
4. `RQ_Free()` removes pending jobs from the queue. it also calls uv_close() which gurentees that `rqAsyncCb` won't be futher invoked, However, the return of uv_close() doesn't ensure the execution of all UV requests, including close_cb(RQ_FreeInternal).
5. `MR_Destroy()` calls MRClust_Free() to release cluster_g resources.
6. `MR_Destroy()` changes `order_for_debug` to allow the background callback continue and call `TrieMap_Delete(currentNodes, (char *)node->id, strlen(node->id), NULL)`. current node and node are have already been released, but since MR_Destroy() didn't reach to the line where it nullifies g_cluster, we are not crashing.
7. the bg callback let the MR_Destroy() (main thread) proceed
8. RQ_Done() sleeps before it tries accesing the queue pointer, since it was nullified in RQ_FREE() and this may lead to a crash.

The next commint willocus on addressing and implementing the fix for this race condition.