Fix MOD-5850, fix crash on tls initialization failure.
The PR fixes a crash on hiredis where the SSL initialization fails. It seems that hiredis do not unset the callbacks that was set by the redisInitiateSSL in case of a failure, and so later those callbacks might be called when the SSL context is not initialized and cause a crash.

An issue was open upstream to fix it on hiredis itself: https://github.com/redis/hiredis/issues/1233

A temporary fix is to reset the callbacks in case of ssl initialization failure.