remove thcount_lock
Thcount used to lock the code block in thread_do where we realize the state is THPOOL_TERMINATE_WHEN_EMPTY and the jobq, and want to inform the other threads. However, this lock was not enough to ensure that thread won't go to sleep after the signal has already been sent.

for example, once the queue is empty,
for thread1 both state = THPOOL_TERMINATE_WHEN_EMPTY and jobq is empty,
switch to thread2: try to pull a job from the queue, is_empty & should run both true, back to thread 1: change should_run and broadcast all other threads
switch to thread2: sleep on the cond

we need to make sure that both code block are protected by the same mutex:
1. thread_do changes the thpool state and signal the other threads
2. jobq_pull check if the jobq is empty and should_run and sleep on the cond.

we lock these code blocks with jobqueues_rwmutex.
thcount_lock is now redundant and the variable it used to guard can be atomic instead.