Add barrier wrapper:
The results are undefined if pthread_barrier_destroy() is called when any thread is blocked on the barrier (that is, has not returned from the pthread_barrier_wait() call).

To guarantee we call  pthread_barrier_destroy() after all threads have returned yet from  pthread_barrier_wait(), i created a wrapper to the barrier, that includes a function that waits for the threads to return and safely destroyed the barrier.

added a test TestWaitTerminate to immitate the behaviour on RDB loading event.