Introduce a global counter for indexing failures for each field type. The counter is update at runtime, once the error occured. It reflects the total number of indexing errors in all living indices. The counter should be decreased whenever an index is dropped.
introduce FieldSpec_AddError that adds an error to a specific field.
it also updates the global field error counter.

We expose the global counter value in INFO modules, along with each field type stats.

the name of the fields count for each type was changed from "Field_type_name" to "Total"