[MOD-8035] Add field index error for JSON field indexing failures (#5150)
* move fieldsStats from RSGlobalConfig to RSGlobalStats

move  dialects stats to global_stats

* Introduce a global counter for indexing failures for each field type.
The counter is update at runtime, once the error occured.
It reflects the total number of indexing errors in all living indices.
The counter should be decreased whenever an index is dropped.

introduce FieldSpec_AddError that adds an error to a specific field.
it also updates the global field error counter.

We expose the global counter value in INFO modules, along with each field type stats.

the name of the fields count for each type was changed from "Field_type_name" to "Total"

* fix test

* skip test_redis_info_errors in cluster

wait for bg index after FT.ALTER to avoid flakyness

* revert change in INFO MODULES field count name

* fix test format

* add field indesing failure when encountered in json

* skip json tests of not json