revert getting empty result
fix bug 1:
aggregate_exec:sendChunk_Resp3
Don't change rc upon timeout, to avoid returning empty results.

bug fix2:
reset sorter timeout so we retrun rof after depleting the heap.

bug fix3:
if shards didn't send any reply (curerntly assuming only possible due to timeout), but they did finish collecting the results, than inprocess would set to 0, but pending is still larger than 0, and the cursor is not depleted.
in this case, we unblock the chanel, signle it to proceed. we rely on the coordinator to check the timeout. If it's timeout it will return.

to allow that, get next reply returns pending (number of shard that potentially have results) instead of 0. in this case rpnet won't return EOF, but timeout.

ehancement:
rpnetNext
cache results len instead of getting it each