Bug fixes related to Timeout with return partial results policy (#5556)
* reset self->timedOut
add test

* revert otiginal test

* suuport empty results (to bereverted)

* revert getting empty result

fix bug 1:
aggregate_exec:sendChunk_Resp3
Don't change rc upon timeout, to avoid returning empty results.

bug fix2:
reset sorter timeout so we retrun rof after depleting the heap.

bug fix3:
if shards didn't send any reply (curerntly assuming only possible due to timeout), but they did finish collecting the results, than inprocess would set to 0, but pending is still larger than 0, and the cursor is not depleted.
in this case, we unblock the chanel, signle it to proceed. we rely on the coordinator to check the timeout. If it's timeout it will return.

to allow that, get next reply returns pending (number of shard that potentially have results) instead of 0. in this case rpnet won't return EOF, but timeout.

ehancement:
rpnetNext
cache results len instead of getting it each

* revert result len

* revert resp2 reply length

* fix rppagerNext_Limit counter update

before this fix self->remaining was decreased and than base->upstream->Next(base->upstream, r) was returned
but the ->Next call could have returned no results, so the counter was wrongly updated for it.
This caused wrong number of accumelated results retuned by cursor, when timeout was reached in one or more of the cursor's read command. In each timedout read we lost one result.

* add optional paramter: [INTERNAL_ONLY]
that will add the timeout Rp only if the query runs in a shard

add tst_timeout
add to debug tests:
TeststrictPolicy
testInternalOnly

add test_cursors:testTimeoutPartialWithEmptyResults

rename parseDebugParams->parseDebugParamsCount

* finalize tests

* fix comment

* disable testTimeoutPartialWithEmptyResults test

* enforce timeout when calling INTERNAL_ONLY with TIMEOUT_AFTER_N where N==0 to avoid coordinator infinite loop

fix missing return in debug commands

* fix tests: explictly settimeout policy to on_timeout return

* fix use after free in rmr.c:MRIteratorCallback_ProcessDone

the problem was that we were trying to decrese &ctx->it->ctx.inProcess when the shard decided to release the iterator (MRIterator_Release released it)
The solution is to change MRIterator_Release signature. Now it returns weather it freed the iterator or not. if it was released, return

* return false if free flag is set to false

* move reader reset of the freeFlag to the uv thread to avoid race conditions with the shards.

ensure the sequence of returning  from MRChannel_Popand re - blocking the chanel
is atomic under the chanel guard

remove MRChannel_Block

* fix leak

* simulate timeout during build - TO BE REVERTED as currently doesn't simulate well the scenario:
i.e -the shard returns timeout error string, butthe coordinator clock is not set to be timeout.
Trying to only return an error from the shards without any special changes in the coordinator: RPNetNext returns error (and exits the while loop) only if the reply contains timeout AND the policy is to fail. otherwise, we enter another while loop iteration without processing the reply, overriding it with the next loop. This is a leak.
Another approach i took was: in the callback passed to the resultprocessor I tried to first call getnextreply before rpnetNext checks the timeout, but it didn't work since for some reason i coudnlt figure (the shards didn't shut down. mught be a bug in RLTest)

Returning a generic query code soen't simulate the scenario required.

* REVERT tieout on buld

* enable test_cursors:testTimeoutPartialWithEmptyResults

skip testEmptyResult in coord

* fix comment

revmove sortable

* stable test_error_with_partial_results by using query debug

* move reader reset of the freeFlag to the uv thread to avoid race conditions with the shards.

ensure the sequence of returning  from MRChannel_Popand re - blocking the chanel
is atomic under the chanel guard

remove MRChannel_Block

* signal one reader

simplify testCursorDepletionNonStrictTimeoutPolicySortby test

* fix typo

* revert force failure