Allow for specifying iteration mode upin TrieMapIterator initialization
We expose the tm_iter_mode as defined in deps/triemap/triemap.h, as well
as a constant `TM_ITER_MODE_DEFAULT` representing the default iteration
mode. TrieMap_Iterate now accepts the iter_mode as the last parameter,
thereby eliminating the need for code external to this crate to modify
`TrieMapIterator`s directly and thereby removing the need for exposing
its internals.