Define TrieMapResultBuf to store results of TrieMap_FindPrefixes
The original TrieMap API defined in `deps/triemap/triemap.h`
specifies the `TrieMap_FindPrefixes` function allocate a buffer
for the results that is to be freed by the caller.

To keep Rust code responsible for both allocating and freeing the result
buffer, we introduce an opaque type `TrieMapResultBuf` that holds the
results of calls to `TrieMap_FindPrefixes`.

Furthermore, this commit introduces the `TrieMapResultBuf_Data`
method to obtain a pointer to the start of the result data buffer,
as well as `TrieMapResultBuf_Free` to free the `TrieMapResultBuf`.