Introduce `LowMemoryThinVec`, a low-memory port of `thin_vec::ThinVec`
`LowMemoryThinVec` is pointer-sized, since it stores its length and capacity
on the heap, in the same allocation used to store its elements.

The implementation is based on a fork of the `thin_vec` crate.
In `low_memory_thin_vec`:
- We remove code paths that are unnecessary for our usecase (e.g.
  `gecko-ffi`, draining, splicing)
- We reduce the maximum capacity of the `std` variant to `u16`,
  matching what we expect to use in our Rust trie re-implementation.
- We leverage newer `std` APIs to simplify existing unsafe blocks
  (e.g. layout computations) and move computations to compile-time
  via `const` annotations where possible.

A fork will allow us to evolve the code in whatever direction
we deem optimal as the porting work progresses.