MOD-9026: Enable debug assertions and overflow checks for Rust tests (#5750)
The `task-test` workflow in CI performs both the build and the test
step using the release profile.
In Rust, this maps to the default `release` profile, which is designed
to produce releaseable artefacts. In particular, it disables debug
assertions as well as overflow checks.

When running tests, we want those safeguards to be enabled.
We therefore define a new profile, named `optimised_test`, which
inherits from the default `release` profile while enabling the
checks we care about and ensure we get nice stacktraces should
a panic occur.

We expose a new knob in `Makefile` to configure the profile of
the Rust portion of the build and set it to `optimised_test` in
the `task_test` CI workflow.