MOD-8970: Define an FFI API for trie_rs corresponding to deps/triemap/triemap.h (#5722)
* Define a FFI API for trie_rs corresponding to deps/triemap/triemap.h

The Rust FFI API exposes TrieMap, TrieMapIterator and TrieMapNode as
opaque types. This removes the ability to update internals of these
types, which in general doesn't happen, except for when the iteration
mode gets altered in src/query.c:1166,1168,1291. In a later commit,
a means of altering the iteration mode during iterator initaliation
will be added.

The code in ffi.rs is validated by running cbindgen against it,
and validating the resulting header file is equivalent to triemap.h.
Example command:
`cd src/redisearch_rs/trie_rs && cbindgen --output triemap_rs.h --lang C`

* Allow for specifying iteration mode upin TrieMapIterator initialization

We expose the tm_iter_mode as defined in deps/triemap/triemap.h, as well
as a constant `TM_ITER_MODE_DEFAULT` representing the default iteration
mode. TrieMap_Iterate now accepts the iter_mode as the last parameter,
thereby eliminating the need for code external to this crate to modify
`TrieMapIterator`s directly and thereby removing the need for exposing
its internals.

* Define TrieMapResultBuf to store results of TrieMap_FindPrefixes

The original TrieMap API defined in `deps/triemap/triemap.h`
specifies the `TrieMap_FindPrefixes` function allocate a buffer
for the results that is to be freed by the caller.

To keep Rust code responsible for both allocating and freeing the result
buffer, we introduce an opaque type `TrieMapResultBuf` that holds the
results of calls to `TrieMap_FindPrefixes`.

Furthermore, this commit introduces the `TrieMapResultBuf_Data`
method to obtain a pointer to the start of the result data buffer,
as well as `TrieMapResultBuf_Free` to free the `TrieMapResultBuf`.

* Document functinality exposed in triemap_rs::ffi

Add safety comments, add debug assertions to validate
input where possible, and adds and expands behavior description
of items in the module.