Implement LowMemoryThinVec::{mem_usage, prepend_with_slice, from_slice}
Adds methods `mem_usage`, `prepend_with_slice`, and `from_slice`
to `LowMemoryThinVec`.
`mem_usage` returns the size of the size of the heap allocation backing
the `LowMemoryThinVec`, `prepend_with_slice` prepends the held
items with a given slice, and `from_slice` creates a `LowMemoryThinVec`
from a slice, cloning the elements.